created 'footer' directory, moved tsx/scss files
