+++ src/App.tsx
@@ -11,8 +11,8 @@ import CompletionBar from "./components/Extras/CompletionBar";
 import WordsPerMinute, { WordsPerMinuteProps } from "./components/Extras/WordsPerMinute";
 import CapsLockIndicator, { CapsLockIndicatorProps } from "./components/Extras/CapsLockIndicator";
 import AfkDetectedIndicator, { AfkDetectedIndicatorProps } from "./components/Extras/AfkDetectedIndicator";
-import Footer, { FooterProps } from "./components/Footer";
-import KeyTips from "./components/KeyTips";
+import Footer, { FooterProps } from "./components/Footer/Footer";
+import KeyTips from "./components/Footer/KeyTips";
 import ColourPaletteSelector, { ColourPaletteSelectorProps } from "./components/Extras/ColourPaletteSelector";
 import UpdateCssVariable from "./components/HelperComponents/UpdateCssVariable";
 import Header, { HeaderProps } from "./components/Header/Header";
