+++ src/components/Footer/Footer.tsx
@@ -1,11 +1,9 @@
-import "../styles/componentStyles/footer.scss";
-
+import "./footer.scss";
 import React from "react";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faCodeBranch, faPalette } from "@fortawesome/free-solid-svg-icons";
 
 export interface FooterProps {
-    
 	setShowColourPalettes: React.Dispatch<React.SetStateAction<boolean>>,
 	colourPaletteDivRef: React.RefObject<HTMLDivElement>
 }
@@ -25,12 +23,10 @@ const Footer = ({setShowColourPalettes, colourPaletteDivRef}: FooterProps) => {
 				<FontAwesomeIcon icon={faPalette} className="palette-icon icon"/>
 				palette
 			</button>
-
 			<a href="https://github.com/JaariReynolds/react-typing-test" target="_blank" title="github.com/JaariReynolds/react-typing-test" className="footer-item" rel="noopener noreferrer">
 				<FontAwesomeIcon icon={faCodeBranch} className="icon"/>
 				<span>git</span>
-			</a>
-			
+			</a>		
 		</div>
 	);
 };
