+++ src/components/Footer/KeyTips.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import "../styles/componentStyles/key-tips.scss";
+import "./key-tips.scss";
 
 const KeyTips = () => {
 	return (
@@ -9,16 +9,8 @@ const KeyTips = () => {
 				<span> + </span>
 				<span className="key-highlight">enter</span>
 				<span> - reset test</span>
-			</div>
-			{/* <div className="key-tip">
-				<span className="key-highlight">control</span>
-				<span> + </span>
-				<span className="key-highlight">q</span>
-				<span> - palette selector</span>
-			</div> */}
-
-		</div>
-		
+			</div>			
+		</div>		
 	);
 };
