+++ src/components/Footer/key-tips.scss
@@ -1,6 +1,5 @@
-@use "../_constants.scss";
-@use "../_variables.scss";
-@use "../_mixins.scss";
+@use "../../styles/mixins";
+@use "../../styles/constants";
 
 .key-tips {
     @include mixins.general-color-and-opacity-transition();
